+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- Updated supported runtimes for functions. Default python runtime is now 3.13. (#8534)
 - Fixed issue where Secret Manager API was not automatically enabled for functions using secrets. (#8528)
 - Changed artifact registry cleanup policy error to warn for CI/CD workloads #8513
 - Enhance firebase init apphosting to support local source deploys. (#8479)